Conflicts:
	visaApplication/src/main/java/com/mastek/visaApplication/entities/ApplicationForm.java
	visaApplication/src/main/java/com/mastek/visaApplication/services/VisaApplicationServices.java
	visaApplication/src/test/java/com/mastek/visaApplication/MongoTests.java
	visaApplication/src/test/java/com/mastek/visaApplication/VisaApplicationTests.java

+++ visaApplication/src/main/java/com/mastek/visaApplication/entities/PersonalDetails.java
@@ -58,7 +58,7 @@ public class PersonalDetails {
 	
 	//Dependants//
 //CREATE DEPENDANTS PASSPORT NUMBER AND PASSPORT DETAILS//
-	long dependantPassportNo;
+	int dependantPassportNo;
 	String dependantFamilyName;
 	Salutation dependantSalutation;
 	Gender dependantGender;
@@ -321,11 +321,11 @@ public class PersonalDetails {
 		NationalInsuranceNumber = nationalInsuranceNumber;
 	}
 	
-	public long getDependantPassportNo() {
+	public int getDependantPassportNo() {
 		return dependantPassportNo;
 	}
 
-	public void setDependantPassportNo(long dependantPassportNo) {
+	public void setDependantPassportNo(int dependantPassportNo) {
 		this.dependantPassportNo = dependantPassportNo;
 	}
